requirements.txt
cog
txtsplit
torch>=2.2.1
torchaudio>=2.2.1
cached_path
transformers
# mecab-python3==1.0.5
num2words==0.5.12
hf_transfer
# unidic_lite==1.0.8
# unidic==1.1.0
pykakasi==2.2.1
fugashi==1.3.0
g2p_en==2.1.0
anyascii==0.3.2
jamo==0.4.1
gruut==2.2.3
g2pkk>=0.1.1
librosa==0.9.1
pydub==0.25.1
eng_to_ipa==0.0.2
inflect==7.0.0
unidecode==1.3.7
pypinyin==0.50.0
cn2an==0.5.22
jieba==0.42.1
gradio
langid==1.1.6
tqdm
tensorboard==2.16.2
loguru==0.7.2
soundfile
myshell-openvoice @ git+https://github.com/remorses/OpenVoice.git@fixed
